Copy dependent DLLs to example build directories
If libvpl.dll is installed in a system location, it will be used at
runtime over a version in the path. This makes it hard to evaluate
runtime behavior of newly downloaded examples.  This change will place
runtime dependencies of the examples alongside the built code so tests
and post build execution will prefer them over ones in system locations.